+++ src/core/commands/index.ts
@@ -6,6 +6,7 @@ import "./skillCommands.js";
 import "./internalHooksCommand.js";
 import "./memoryCommand.js";
 import "./mcpCommand.js";
+import "./workspaceCommand.js";
 
 export { registry } from "./registry.js";
 export * from "./types.js";